+++ gax/src/main/java/com/google/api/gax/rpc/ApiCallContext.java
@@ -36,6 +36,7 @@ import com.google.api.gax.retrying.RetryingContext;
 import com.google.api.gax.rpc.StatusCode.Code;
 import com.google.api.gax.tracing.ApiTracer;
 import com.google.auth.Credentials;
+import com.google.common.base.Preconditions;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
@@ -242,4 +243,31 @@ public interface ApiCallContext extends RetryingContext {
   /** Return the extra headers set for this context. */
   @BetaApi("The surface for extra headers is not stable yet and may change in the future.")
   Map<String, List<String>> getExtraHeaders();
+
+  /**
+   * Return a new ApiCallContext with additional option merged into the present instance. Any
+   * existing value of the key is overwritten.
+   */
+  @BetaApi("The surface for call context options is not stable yet and may change in the future.")
+  <T> ApiCallContext withOption(Key<T> key, T value);
+
+  /** Return the api call context option set for this context. */
+  @SuppressWarnings("unchecked")
+  @BetaApi("The surface for call context options is not stable yet and may change in the future.")
+  <T> T getOption(Key<T> key);
+
+  /** Key for api call context options key-value pair. */
+  final class Key<T> {
+    private final String name;
+
+    private Key(String name) {
+      this.name = name;
+    }
+
+    /** Factory method for creating instances of {@link Key}. */
+    public static <T> Key<T> create(String name) {
+      Preconditions.checkNotNull(name, "Key name cannot be null.");
+      return new Key<>(name);
+    }
+  }
 }